+++ packages/framework/tests/Unit/HydeBasePathCanBeChangedTest.php
@@ -7,11 +7,9 @@ namespace Hyde\Framework\Testing\Unit;
 use Hyde\Hyde;
 use Hyde\Testing\UnitTestCase;
 
-/**
- * @covers \Hyde\Foundation\HydeKernel::getBasePath
- * @covers \Hyde\Foundation\HydeKernel::setBasePath
- * @covers \Hyde\Foundation\HydeKernel::path
- */
+#[\PHPUnit\Framework\Attributes\CoversClass(\Hyde\Foundation\HydeKernel::getBasePath)]
+#[\PHPUnit\Framework\Attributes\CoversClass(\Hyde\Foundation\HydeKernel::setBasePath)]
+#[\PHPUnit\Framework\Attributes\CoversClass(\Hyde\Foundation\HydeKernel::path)]
 class HydeBasePathCanBeChangedTest extends UnitTestCase
 {
     protected static bool $needsKernel = true;