+++ packages/framework/tests/Unit/Views/BlogPostFeedComponentViewTest.php
@@ -9,9 +9,7 @@ use Hyde\Testing\TestCase;
 use Hyde\Pages\MarkdownPost;
 use Hyde\Testing\TestsBladeViews;
 
-/**
- * @coversNothing Test to ensure the blog post feed component can be rendered
- */
+#[\PHPUnit\Framework\Attributes\CoversNothing]
 class BlogPostFeedComponentViewTest extends TestCase
 {
     use TestsBladeViews;